Extract createConstraintsSlice from projectStore (P5 store slice)
