+++ src/store/INDEX.md
@@ -23,6 +23,7 @@ Zustand store. The single source of truth for the current `.camj` project. **All
   - `machineDefsSlice.ts` — machine definition CRUD: set selected, add/remove/refresh machine definitions
   - `operationsSlice.ts` — operation CRUD, rest-operation creation, toolpath visibility, duplication, and ordering
   - `importMergeSlice.ts` — shape import and `.camj` folder merge actions
+  - `constraintsSlice.ts` — persistent fixed-distance constraint placement, value updates, cancellation, and deletion
 - `helpers/` — pure helpers used by the store
   - `clipping.ts` — clipper-lib wrappers (handles the integer scaling factor): profile↔Clipper-path conversion, boolean/offset execution, and overlap predicates. Arc/curve reconstruction of Clipper output lives in `engine/toolpaths/arcReconstruction.ts`.
   - `derivedFeatures.ts` — computes derived features from the feature tree
